[CLI-47] Fix bug: 'virgil identity-confirm-global' doesn't work
Identity confirmation with requests number and time validity limit results in an error "--action-id and
--confirmation-code must always together"

Description
1. Confirm an identity with requests number limit = 10 and time validity limit = 60:
virgil identity-confirm-global -d email:alice@gmail.com -o validated-identity.txt -t 60 -c 10
2. Enter the confirmation code you've received on alice@gmail.com.

*Actual Result*
identity-confirm-global. Error: --action-id and --confirmation-code must always together
The message for entering the confirmation code is "Enter confirmation code which was sent on you identity"

*Expected Result*
The identity is confirmed.
The message for entering the confirmation code is "Enter the confirmation code that was sent on" or "Enter the
confirmation code that was sent on your email".